requirements.txt
feedparser>=6.0.10
GitPython>=3.1.40